+++ src/Development/Shake/Internal/Unit.hs
@@ -1,8 +1,18 @@
-{-# LANGUAGE CPP #-}
-module Development.Shake.Internal.Unit (Unit (..)) where
+{-# LANGUAGE CPP                  #-}
+{-# LANGUAGE FlexibleInstances    #-}
+{-# LANGUAGE GADTs                #-}
+#if __GLASGOW_HASKELL__ < 710
+{-# LANGUAGE OverlappingInstances #-}
+#endif
+module Development.Shake.Internal.Unit (Unit) where
+
 #if __GLASGOW_HASKELL__ >= 710
-import Development.Shake.Internal.UnitNewGHC
+class Unit a
+instance {-# OVERLAPPING #-} Unit b => Unit (a -> b)
+instance {-# OVERLAPPABLE #-} a ~ () => Unit (m a)
 #else
-import Development.Shake.Internal.UnitOldGHC
+class Unit a
+instance Unit b => Unit (a -> b)
+instance a ~ () => Unit (m a)
 #endif
