--- src/Development/Shake/Internal/UnitOldGHC.hs
@@ -1,8 +1,0 @@
-{-# LANGUAGE FlexibleInstances    #-}
-{-# LANGUAGE GADTs                #-}
-{-# LANGUAGE OverlappingInstances #-}
-module Development.Shake.Internal.UnitOldGHC (Unit) where
-
-class Unit a
-instance Unit b => Unit (a -> b)
-instance a ~ () => Unit (m a)
